# src/main/resources/JacoboMil-io-builders-2.0.0-oas3-resolved.yaml
openapi: 3.0.0
info:
  description: IOBuilders-Bank
  version: "1.0.0"
  title: This is a Rest API for ioBuilders-bank
  contact:
    email: you@your-company.com
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
servers:
  - url: /v1
...
paths:
...
  /users/{userId}:
    get:
      tags:
      - Users
      summary: Get user info
      description: Retrieve user information
      operationId: get_user_by_id
      parameters:
      - name: userId
        in: path
        required: true
        style: simple
        explode: false
        schema:
          maxLength: 36
          pattern: \b[0-9a-f]{8}\b-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-\b[0-9a-f]{12}\b
          type: string
          format: uuid
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
    delete:
      tags:
        - Users
      summary: Delete user
      description: Delete user information
      operationId: delete_user
      parameters:
        - name: userId
          in: path
          required: true
          style: simple
          explode: false
          schema:
            maxLength: 36
            pattern: \b[0-9a-f]{8}\b-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-\b[0-9a-f]{12}\b
            type: string
            format: uuid
      responses:
        204:
          description: No content
...
components:
  schemas:
...
    UserResponse:
      type: object
      properties:
        userId:
          maxLength: 36
          pattern: \b[0-9a-f]{8}\b-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-\b[0-9a-f]{12}\b
          type: string
          format: uuid
          example: xxxxxxxx-xxxx-xxxx-xxxx-xxxxxxxxxxxx
        username:
          maxLength: 50
          pattern: /^[a-z ,.'-]+$/i
          type: string
          example: johndoe
        firstName:
          maxLength: 50
          pattern: /^[a-z ,.'-]+$/i
          type: string
          example: John
        lastName:
          maxLength: 50
          pattern: /^[a-z ,.'-]+$/i
          type: string
          example: Doe
        email:
          pattern: (?:[a-z0-9!#$%&'*+/=?^_`{|}~-]+(?:\.[a-z0-9!#$%&'*+/=?^_`{|}~-]+)*|"(?:[\x01-\x08\x0b\x0c\x0e-\x1f\x21\x23-\x5b\x5d-\x7f]|\\[\x01-\x09\x0b\x0c\x0e-\x7f])*")@(?:(?:[a-z0-9](?:[a-z0-9-]*[a-z0-9])?\.)+[a-z0-9](?:[a-z0-9-]*[a-z0-9])?|\[(?:(?:25[0-5]|2[0-4][0-9]|[01]?[0-9][0-9]?)\.){3}(?:25[0-5]|2[0-4][0-9]|[01]?[0-9][0-9]?|[a-z0-9-]*[a-z0-9]:(?:[\x01-\x08\x0b\x0c\x0e-\x1f\x21-\x5a\x53-\x7f]|\\[\x01-\x09\x0b\x0c\x0e-\x7f])+)\])
          type: string
          example: jdoe@iobuilders.com
...